+++ packages/docs/.storybook/main.js
@@ -6,8 +6,10 @@ module.exports = {
   "addons": [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
+    "@storybook/addon-actions",
     "@storybook/addon-interactions",
-    "@storybook/addon-a11y"
+    "@storybook/addon-a11y",
+    "@storybook/client-api"
   ],
   "framework": "@storybook/react",
   core: {