+++ packages/docs/.storybook/preview.js
@@ -2,6 +2,7 @@ import { themes } from "@storybook/theming";
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
+  layout: 'centered',
   controls: {
     matchers: {
       color: /(background|color)$/i,