+++ packages/docs/src/stories/Box.stories.tsx
@@ -1,4 +1,4 @@
-import { Box, BoxProps, Text } from '@lm-ignite/react-ignite-tokens'
+import { Box, BoxProps, Text } from '@lm-ignite/react-ignite'
 import type { StoryObj, Meta } from '@storybook/react'
 
 export default {