+++ packages/docs/src/stories/Checkbox.stories.tsx
@@ -3,7 +3,7 @@ import {
   Text,
   Checkbox,
   CheckboxProps,
-} from '@lm-ignite/react-ignite-tokens'
+} from '@lm-ignite/react-ignite'
 import type { StoryObj, Meta } from '@storybook/react'
 
 export default {