+++ packages/react/src/components/Text.tsx
@@ -1,5 +1,5 @@
 import { ComponentProps } from '@stitches/react'
-import { styled } from '../styles'
+import { styled } from '../styles/stitches.config'
 
 import { fontSizes } from '@lm-ignite/ignite-tokens'
 import { ElementType } from 'react'