+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lm-ignite/eslint-config
 
+## 1.3.0
+
+### Minor Changes
+
+- Sending all changes
+
+## 1.2.0
+
+### Minor Changes
+
+- Finishing ignite challenge
+
 ## 1.1.0
 
 ### Minor Changes