+++ packages/react/CHANGELOG.md
@@ -1,4 +1,16 @@
-# @lm-ignite/react-ignite-tokens
+# @lm-ignite/react-ignite
+
+## 2.3.0
+
+### Minor Changes
+
+- Sending all changes
+
+## 2.2.0
+
+### Minor Changes
+
+- Finishing ignite challenge
 
 ## 2.1.0