+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lm-ignite/ignite-tokens
 
+## 2.3.0
+
+### Minor Changes
+
+- Sending all changes
+
+## 2.2.0
+
+### Minor Changes
+
+- Finishing ignite challenge
+
 ## 2.1.0
 
 ### Minor Changes